feat: update nav for friends
